requirements.txt
alembic==1.4.0
aniso8601==8.0.0
bcrypt==3.1.7
cffi==1.14.0
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-JWT==0.3.2
Flask-Migrate==2.5.2
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.4.2
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
validate-email==1.3
Werkzeug==1.0.0
